refactor:useComment 훅 사용하여 Comment컴포넌트 코드 간소화
