+++ src/components/Community/WriteBoardMain.jsx
@@ -39,7 +39,7 @@ export default function WriteBoardMain({ post }) {
   return (
     <>
       <Container>
-        <RowContainer >
+        <RowContainer>
           <ProfileImg
             src={post.profileUrl || profile}
             alt="Profile"
@@ -78,14 +78,11 @@ export default function WriteBoardMain({ post }) {
         </HeartContainer>
       </Container>
       <RowLine />
-      <Container>
-        <Comment post={post} />
-      </Container>
+      <Comment post={post} />
     </>
   );
 }
 const Container = styled.div`
-
   padding: 0px 20px;
 `;
 const ProfileImg = styled.img`
