+++ src/hooks/useComment.jsx
@@ -1,62 +1,51 @@
 // hooks/useComments.js
-import { useState, useCallback, useEffect } from "react";
+import { useState, useEffect } from "react";
 import { getCommentApi } from "../apis/communityApi/getCommentApi";
 import { commentApi } from "../apis/communityApi/commentApi";
 
-export const useComments = (postId) => {
-  // 댓글  상태 관리
+export const useComment = (post) => {
+  // 해당 게시글 댓글 불러오기
   const [comments, setComments] = useState([]);
-  const [page, setPage] = useState(1);
-  const [hasMore, setHasMore] = useState(true);
+  // 무한 스크롤 시 로딩
   const [isLoading, setIsLoading] = useState(false);
+  // 더 불러올수 있는지 여부
+  const [hasMore, setHasMore] = useState(true);
+  // 댓글 5개씩 페이지네이션 페이지번호
+  const [page, setPage] = useState(1);
 
-  // 댓글을 불러오는 함수
-  const fetchComments = useCallback(async () => {
-    if (isLoading || !hasMore) return;
-    setIsLoading(true);
+  // 각 댓글 관리
+  const [comment, setComment] = useState("");
+  const isCommentEmpty = comment.trim().length === 0;
+
+  const handleCommentSubmit = async () => {
+    // 빈 댓글 작성 시 버튼 클릭 금지 및 비활성화
+    if (isCommentEmpty) return;
+
     try {
-      // API를 호출하여 댓글 불러오기
-      await getCommentApi(postId, page, setComments, setIsLoading, setHasMore);
-      setPage((prevPage) => prevPage + 1); // 다음 페이지로 이동
+      //새 댓글 작성
+      await commentApi(post.id, comment);
+      // 입력칸 초기화
+      setComment("");
+
+      setPage(1);
+      //댓글 배열 초기화
+      setComments([]);
+      //새 댓글이 작성된 새로운 댓글 목록 불러오기.
+      await getCommentApi(post.id, 1, setComments, setIsLoading, setHasMore);
     } catch (error) {
-      console.error("댓글을 불러오는 중 오류가 발생했습니다:", error);
-    } finally {
-      setIsLoading(false);
+      console.log("댓글 전송 실패", error);
     }
-  }, [postId, page, hasMore, isLoading]);
-
-  // 새 댓글을 추가
-  const addComment = useCallback(
-    async (content) => {
-      try {
-        // API를 호출하여 새 댓글 추가
-        const newComment = await commentApi(postId, content);
-        setComments((prevComments) => [newComment, ...prevComments]); // 새 댓글을 목록 맨 앞에 추가
-
-        setPage(1);
-        setComments([]);
-        fetchComments();
-      } catch (error) {
-        console.error("댓글 추가 중 오류가 발생했습니다:", error);
-      }
-    },
-    [postId, fetchComments]
-  );
-
-  // postId가 변경될 때마다 댓글 목록 초기화 및 새로 불러오기
-  useEffect(() => {
-    setPage(1);
-    setComments([]);
-    setHasMore(true);
-    fetchComments();
-  }, [postId]);
-
-  
-  return {
-    comments,
-    isLoading,
-    hasMore,
-    fetchComments,
-    addComment,
   };
+
+  useEffect(() => {
+    const fetchComments = async () => {
+      setComments([]);
+      setPage(1);
+      setHasMore(true);
+      await getCommentApi(post.id, 1, setComments, setIsLoading, setHasMore);
+    };
+    fetchComments();
+  }, [post.id]);
+
+  return {comments,comment,setComment,handleCommentSubmit,isCommentEmpty};
 };
